+++ packages/opa/RELEASES.md
@@ -7,3 +7,11 @@ Based on:
 - Speakeasy CLI 1.322.1 (2.354.2) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [typescript v1.1.3] packages/opa
+
+## 2024-07-01 07:54:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.322.3 (2.354.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.1.4] packages/opa